==> Notebooks/nb8.2_lumo_ipes_predictor.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import csv

# Open the CSV file for reading
with open('outputs/lumo_optimum_features_and_trials.csv', 'r', newline='') as file:
    reader = csv.reader(file)

    # Read the first row from the CSV file
    row = next(reader)

    # Assign the values to a and b
    optimum_number_of_features, optimum_number_of_trials = int(row[0]), int(row[1])

# Now a and b contain the values from the CSV file
print(f"Number of features: {optimum_number_of_features}, Number of trials: {optimum_number_of_trials}")

## Target dataset prep

import pickle
with open('outputs/non_constant_columns.pkl', 'rb') as f:
    non_constant_columns = pickle.load(f)

target_df = pd.read_pickle('outputs/target_descriptors_calculated_n_processed.pkl')
target_df.head()

target_df.info()

# Only use the molecules with a LUMO_IPES value
target_df = target_df[target_df['LUMO_IPES'].notna()]

from sklearn.utils import shuffle
target_df  = shuffle(target_df, random_state=0)
target_df.head()

X = target_df[target_df['Type'] != 'External Validation'][non_constant_columns]
y = target_df[target_df['Type'] != 'External Validation']['LUMO_IPES']

X.shape, y.shape

from sklearn.model_selection import KFold
from sklearn.metrics import mean_absolute_error
from sklearn.preprocessing import StandardScaler

from sklearn.feature_selection import RFE
from sklearn.ensemble import RandomForestRegressor

rf = RandomForestRegressor(random_state=0)
selector = RFE(rf, n_features_to_select=optimum_number_of_features-1, step=1)
selector = selector.fit(X, y)

X_selected = selector.transform(X)
X_selected.shape

selected_features = X.columns[selector.support_]
selected_features

X_selected = X[X.columns[selector.support_]]
X_selected

## Combine selected features with DFT prediction

from joblib import dump, load
base_model = load('outputs/models/lumo_dft.joblib')

base_model_preds = base_model.predict(X)
base_model_preds = base_model_preds.reshape(-1, 1)

X_selected.insert(0, 'DFT_pred', base_model_preds)

selected_features = list(X_selected.columns)
selected_features

type(X_selected), type(y)

from sklearn.model_selection import RepeatedKFold
from sklearn.model_selection import cross_validate
cv = RepeatedKFold(n_splits=10, n_repeats=5, random_state=0)
rf = RandomForestRegressor(random_state=0)
cv_results = cross_validate(rf, X_selected, y, scoring='neg_mean_absolute_error', return_estimator=True, cv=cv)
np.mean(cv_results['test_score'])*-1

import optuna
from sklearn.ensemble import RandomForestRegressor
from sklearn.model_selection import RepeatedKFold

cv = RepeatedKFold(n_splits=10, n_repeats=5, random_state=0)

def objective(trial):
    params = {
        'n_estimators': trial.suggest_int('n_estimators', 10, 700),
        'max_depth': trial.suggest_int('max_depth', 2, 25),
        'min_samples_split': trial.suggest_int('min_samples_split', 2, 10),
        'min_samples_leaf': trial.suggest_int('min_samples_leaf', 1, 5),
        'max_features': trial.suggest_float('max_features', 0.3, 1.0)
        }

    rf = RandomForestRegressor(random_state=0, **params)
    cv_results = cross_validate(
        rf,
        X_selected, y,
        scoring='neg_mean_absolute_error',
        return_estimator=True,
        cv=cv,
        n_jobs=-1)
    return np.mean(cv_results['test_score'])*-1

study = optuna.create_study(direction='minimize', sampler=optuna.samplers.TPESampler(seed=0))
study.optimize(objective, n_trials=optimum_number_of_trials)

study.best_params

study.best_value

cv = RepeatedKFold(n_splits=10, n_repeats=5, random_state=0)
rf = RandomForestRegressor(random_state=0, **study.best_params)
cv_results = cross_validate(rf, X_selected, y, scoring='neg_mean_absolute_error', return_estimator=True, cv=cv)
np.mean(cv_results['test_score'])*-1

from sklearn.model_selection import cross_val_predict, KFold
from sklearn.ensemble import RandomForestRegressor
errors = []
for random_state in range(5):
    cv = KFold(n_splits=10, shuffle=True, random_state=random_state)
    preds = cross_val_predict(rf, X_selected, y, cv=cv)
    errors.append(np.abs(preds-y))

errors_per_molecule = np.mean(np.array(errors), axis=0)

result = target_df[['Type', 'Molecule', 'LUMO_IPES']]
result = result[result['Type'] != 'External Validation']
result['Error'] = errors_per_molecule
result.sort_values(by='Error')

np.mean(result['Error'][result['Type'] == 'Target Dataset'])

np.mean(result['Error'][result['Type'] == 'Online Dataset'])

## Final model

lumo_ipes_predictor = RandomForestRegressor(random_state=0, **study.best_params)
lumo_ipes_predictor.fit(X_selected, y)

from joblib import dump, load
dump(lumo_ipes_predictor, 'outputs/models/lumo_ipes_predictor.joblib')

# Feature Importance
df = pd.DataFrame({
    "Feature Name": lumo_ipes_predictor.feature_names_in_,
    "Feature Importance": lumo_ipes_predictor.feature_importances_
})

# Sort the DataFrame by 'Feature Importance' in descending order
df_sorted = df.sort_values(by="Feature Importance", ascending=False)

# Display the sorted DataFrame
df_sorted
