==> Notebooks/nb9.2_external_validation_lumo.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import pickle
with open('outputs/non_constant_columns.pkl', 'rb') as f:
    non_constant_columns = pickle.load(f)

target_df = pd.read_pickle('outputs/target_descriptors_calculated_n_processed.pkl')
target_df.head()

target_df[target_df['Type'] == 'External Validation']

X_test = target_df[target_df['Type'] == 'External Validation'][non_constant_columns]
y_test = target_df[target_df['Type'] == 'External Validation']['LUMO_IPES']

X_test.shape, y_test.shape

from joblib import dump, load
base_model = load('outputs/models/lumo_dft.joblib')

from joblib import dump, load
lumo_ipes_predictor = load('outputs/models/lumo_ipes_predictor.joblib')

lumo_dft_predictions = base_model.predict(X_test)
lumo_dft_predictions

# the features names random forest trained on
lumo_ipes_predictor.feature_names_in_

descriptors_used_for_training = list(lumo_ipes_predictor.feature_names_in_)[1:]
descriptors_used_for_training

X_selected = X_test[descriptors_used_for_training]
X_selected.shape

X_selected.insert(0, 'DFT_pred', lumo_dft_predictions)

X_selected

preds = lumo_ipes_predictor.predict(X_selected)

results_df = target_df[target_df['Type'] == 'External Validation']
results_df = results_df[['Molecule', 'LUMO_IPES']]
results_df['LUMO_IPES_prediction'] = preds
results_df['Error'] = np.abs(results_df['LUMO_IPES_prediction'] - results_df['LUMO_IPES'])
results_df

# Mean Absolute Error
np.mean(results_df['Error'])
